test fixed and secret handled
