+++ .github/workflows/ci.yml
@@ -32,6 +32,10 @@ jobs:
           DATA: ${{ secrets.GOOGLE_SERVICES_JSON }}
         run: echo $DATA | base64 -di > app/google-services.json
 
+      # Handle secret file
+      - name: Load secrets.properties
+        run: echo "${{secrets.SECRETS_PROPERTIES }}" > $GITHUB_WORKSPACE/secrets.properties
+
       # Install the Firebase CLI
       - name: Install firebase
         run: curl -sL https://firebase.tools | bash
