+++ secrets.properties
@@ -1,0 +1,1 @@
+GOOGLE_MAPS_API_KEY = AIzaSyAHkGMX1AGthHeTC7LU8sKfyEAJ2danMgo
